feat: 카드 이미지 클릭시 모달 뜨는 기능 추가
