+++ react-payments-app/src/components/domain/Modal/CompanySelectionModal.jsx
@@ -1,0 +1,48 @@
+const CompanySelectionModal = () => {
+  return (
+    <div>
+      <div className='modal-dimmed'>
+        <div className='modal'>
+          <div className='flex-center'>
+            <div className='modal-item-container'>
+              <div className='modal-item-dot'></div>
+              <span className='modal-item-name'>클린 카드</span>
+            </div>
+            <div className='modal-item-container'>
+              <div className='modal-item-dot'></div>
+              <span className='modal-item-name'>클린 카드</span>
+            </div>
+            <div className='modal-item-container'>
+              <div className='modal-item-dot'></div>
+              <span className='modal-item-name'>클린 카드</span>
+            </div>
+            <div className='modal-item-container'>
+              <div className='modal-item-dot'></div>
+              <span className='modal-item-name'>클린 카드</span>
+            </div>
+          </div>
+          <div className='flex-center'>
+            <div className='modal-item-container'>
+              <div className='modal-item-dot'></div>
+              <span className='modal-item-name'>클린 카드</span>
+            </div>
+            <div className='modal-item-container'>
+              <div className='modal-item-dot'></div>
+              <span className='modal-item-name'>클린 카드</span>
+            </div>
+            <div className='modal-item-container'>
+              <div className='modal-item-dot'></div>
+              <span className='modal-item-name'>클린 카드</span>
+            </div>
+            <div className='modal-item-container'>
+              <div className='modal-item-dot'></div>
+              <span className='modal-item-name'>클린 카드</span>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default CompanySelectionModal;
