+++ react-payments-app/src/pages/CardRegistrationPage/CardRegistrationPage.jsx
@@ -3,66 +3,22 @@ import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import { useCard } from '../../store/CardContext';
-import { MAX_INPUT_LENGTH } from '../../constants/numbers';
 
 import Header from '../../components/common/Header/Header';
 import Card from '../../components/domain/Card/Card';
 import CardForm from '../../components/domain/CardForm/CardForm';
+import CompanySelectionModal from '../../components/domain/Modal/CompanySelectionModal';
 
 const CardRegistration = () => {
   const navigate = useNavigate();
   const { cardInfo } = useCard();
-  const [cardNumbers, setCardNumbers] = useState([]);
-  const [cardExpirationDate, setCardExpirationDate] = useState([]);
-  const [cardOwner, setCardOwner] = useState([]);
-  const [cardCVC, setCardCVC] = useState([]);
-  const [cardPassword, setCardPassword] = useState([]);
+
+  const [isModalOpen, setIsModalOpen] = useState(false);
 
   const goBackToListPage = (e) => {
     navigate('/');
   };
 
-  const isAllFilledOut = (cardInfo) => {
-    const { cardNumbers, cardExpirationDate, cardPassword, cardCVC } = cardInfo;
-
-    return (
-      cardCVC.length === MAX_INPUT_LENGTH.CVC &&
-      cardNumbers.every(
-        (part) => part.length === MAX_INPUT_LENGTH.CARD_NUMBER
-      ) &&
-      cardExpirationDate.every(
-        (part) => part.length === MAX_INPUT_LENGTH.DATE
-      ) &&
-      cardPassword.every((part) => part.length === MAX_INPUT_LENGTH.PW)
-    );
-  };
-
-  const handleSubmit = (cardInfo) => {
-    // if (!isAllFilledOut(cardInfo)) {
-    //   alert('빈칸을 모두 채워주세요.');
-    //   return;
-    // }
-    // setCard(cardInfo);
-    // navigate('/registration/setCardNickname');
-  };
-
-  const handleChange = (inputInfo) => {
-    const {
-      cardNumbers,
-      cardExpirationDate,
-      cardOwner,
-      cardCVC,
-      cardPassword,
-    } = inputInfo;
-
-    if (inputInfo) {
-      setCardNumbers(cardNumbers);
-      setCardExpirationDate(cardExpirationDate);
-      setCardOwner(cardOwner);
-      setCardCVC(cardCVC);
-      setCardPassword(cardPassword);
-    }
-  };
   return (
     <>
       <Header
@@ -75,8 +31,10 @@ const CardRegistration = () => {
         cardNumbers={cardInfo.cardNumbers}
         cardExpirationDate={cardInfo.cardExpirationDate}
         cardOwner={cardInfo.cardOwner}
+        onClick={() => setIsModalOpen(!isModalOpen)}
       />
-      <CardForm onSubmit={handleSubmit} onChange={handleChange} />
+      <CardForm />
+      {isModalOpen && <CompanySelectionModal />}
     </>
   );
 };
